Move manuscript code to own folder
