+++ novelwriter/manuscript/buildsettings.py
